Working and tested.
The parser is exhaustive and could be made simpler/prioritized to focus
on less things. However, I feel like the ability to parse the additional
information out of Mimikatz's output will prove useful in the future, so
I intend to leave it in, as although expansive, the additional parsing
work is straightforward.